Modify lab initialization so first preview works
Remove unused binding in nodeHistogram
Move joint js interaction to labUtils.service
On first preview, scale breakpoints to histogram before previewing
Fix histogram fetching
