+++ app-frontend/src/app/components/histogram/histogramBreakpoint/histogramBreakpoint.html
@@ -11,7 +11,7 @@
 >
   <input type="number"
          class="breakpoint-input-field"
-         ng-attr-step="{{'any'}}"
+         ng-attr-step="{{$ctrl.precision}}"
          ng-attr-max="{{$ctrl.range.max}}"
          ng-attr-min="{{$ctrl.range.min}}"
          ng-model="$ctrl.breakpoint"
