+++ app-frontend/src/app/components/histogram/nodeHistogram/nodeHistogram.component.js
@@ -7,7 +7,6 @@ const nodeHistogram = {
         histogram: '<',
         breakpoints: '<',
         options: '<',
-        onMasksChange: '&',
         onBreakpointChange: '&'
     }
 };
