+++ app-frontend/src/app/components/tools/diagramContainer/diagramContainer.component.js
@@ -10,6 +10,7 @@ export default {
         onPreview: '&',
         onShare: '&',
         onParameterChange: '&',
-        onGraphComplete: '&'
+        onGraphComplete: '&',
+        createToolRun: '&'
     }
 };
